+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * New "Recent" and "Favorites" sections on [/rpc-browser](./rpc-browser)
 * On blocks lists show min/avg/max fee rates instead of just avg
 * Random Bitcoin-related quote shown in footer on each page load
+* New [/quotes](./quotes), curated list of Bitcoin-related quotes (each quote also having its own page like [this](`./quote/0`))
 * Fix for incorrect homepage block count when using `BTCEXP_UI_HOME_PAGE_LATEST_BLOCKS_COUNT`
 * Fix for inaccurate difficulty adjustment estimates
 * Link to Tor v3 Hidden Service in footer